requirements/base.txt
mongoengine
pydantic-settings
lingua-language-detector